tests/requirements-pre314.txt
-r requirements-minimal.txt
aioamqp>=0.15.0
aiofiles>=0.5.0
aiohttp>=3.8.3
aio-pika>=9.5.2
boto3>=1.17.74
bottle>=0.12.25
celery>=5.2.7
Django>=4.2.16
# FastAPI depends on pydantic-core which requires rust to be installed and
# it's not compiling due to python_version restrictions.
# fastapi>=0.115.0; python_version >= "3.13"
flask>=2.3.2
# gevent is taking more than 20min to build on 3.14
# gevent>=1.4.0
grpcio>=1.14.1
google-cloud-pubsub>=2.0.0
google-cloud-storage>=1.24.0
legacy-cgi>=2.6.1
lxml>=4.9.2
mock>=4.0.3
moto>=4.1.2
mysqlclient>=2.0.3
PyMySQL[rsa]>=1.0.2
psycopg2-binary>=2.8.6
pika>=1.2.0
pymongo>=3.11.4
pyramid>=2.0.1
pytest-mock>=3.12.0
pytz>=2024.1
redis>=3.5.3
requests-mock
responses<=0.17.0
# Sanic is not installable on 3.13 because `httptools, uvloop` dependencies fail to compile:
# `too few arguments to function ‘_PyLong_AsByteArray’`
sanic>=19.9.0; python_version < "3.13"
sanic-testing>=24.6.0; python_version < "3.13"
starlette>=0.38.2
sqlalchemy>=2.0.0
tornado>=6.4.1
uvicorn>=0.13.4
urllib3>=1.26.5
httpx>=0.27.0
protobuf<=6.30.2
